requirements.txt
streamlit
openai
SimplerLLM
lxml
python-dotenv
lxml[html_clean]
requests